Rename expression_provider to _service
